+++ src/frontend/ui/recherche-avancee/resultat-recherche-avancee/ExportExcelRechercheAvancee.tsx
@@ -1,10 +1,13 @@
 
 
+import { sendEvent } from "@socialgouv/matomo-next";
 import { Workbook } from "exceljs";
 import { useContext } from "react";
 
 import { Résultat, RésultatDeRecherche } from "../../../../backend/métier/entities/RésultatDeRecherche";
 import { ecrireLignesDansSheet, telechargerWorkbook } from "../../../utils/excelUtils";
+import { EXPORT } from "../../../utils/nomenclature-matomo";
+import { SourceMatomo } from "../../../utils/SourceMatomo";
 import { RechercheAvanceeContext, RechercheAvanceeContextValue } from "../../commun/contexts/RechercheAvanceeContext";
 import { UserContext } from "../../commun/contexts/userContext";
 import { UserListViewModel } from "../../user-list/UserListViewModel";
@@ -211,6 +214,8 @@ function generateCriteriaData(context: RechercheAvanceeContextValue, categoriesF
 }
 
 async function generateAndExportExcel(context: RechercheAvanceeContextValue | undefined, favoris: UserListViewModel[] | undefined, categories: CategoriesFinessViewModel[]) {
+  sendEvent(EXPORT(SourceMatomo.RECHERCHE_AVANCEE));
+
   if (context) {
     const fileName: string = `${getCurrentDate()}_Helios_rechercheavancee.xlsx`;
     const data = await getData(context);